requirements/base.txt
fastapi==0.95.1                 # FastApi
pydantic==1.10.4                # FastApi dependency
email-validator==1.3.0          # FastApi dependency
uvicorn==0.20.0                 # FastApi dependency
python-multipart==0.0.5         # FastApi dependency
sqlalchemy==1.4.46              # Database / SQL ORM
psycopg2==2.9.5                 # Database / PostgreSQL driver
databases[asyncpg]==0.7.0       # Database / Async PostgreSQL driver
sendgrid==6.9.7                 # Notifications / Email Service
firebase-admin==6.0.1           # Firebase Admin SDK
python-dateutil==2.8.2          # Date utilities
joblib==1.2.0                   # ML / Save and Load Machine Learning Models
scikit-learn==1.2.0             # ML / Machine Learning Library
pandas==1.5.3                   # ML / Data Analysis Library
tensorflow==2.11.0              # ML / Machine Learning Library
Pillow==9.4.0                   # ML / Machine Learning Library
numpy==1.24.3                   # ML / Data Analysis Library
Whoosh==2.7.4                   # Searching
jinja2==3.1.2                   # PDF / HTML to PDF
pdfkit==1.0.0                   # PDF / HTML to PDF
PyPDF2==3.0.1                   # PDF / Merge PDFs


# Testing and Linting. These libraries are not required for the application to run
pytest==7.2.1                   # Testing
pytest-env==0.8.1               # Testing
httpx==0.23.3                   # Testing
flake8==6.0.0                   # Linting
isort==5.11.4                   # Linting
black==23.3.0                   # Linting
python-dotenv==0.21.1           # Load environment variables from .env file, used only in /scripts folder
openpyxl==3.1.2                 # Read and write Excel files, used only in /scripts folder
async-asgi-testclient==1.4.11   # Testing
pytest-asyncio==0.21.0          # Testing
coverage==7.0.3                 # Testing
pytest-cov==4.1.0               # Testing

